add http flag; db-server debug code clean-ups
